% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{detect_packages}
\alias{detect_packages}
\title{Detect R packages in the project with optional versions}
\usage{
detect_packages(
  project_dir,
  include_versions = FALSE,
  use_latest_versions = FALSE
)
}
\arguments{
\item{project_dir}{The path to the project directory}

\item{include_versions}{Logical indicating whether to include package versions}

\item{use_latest_versions}{Logical indicating whether to use the latest versions}
}
\value{
A named vector of package names and their versions (if requested)
}
\description{
Detect R packages in the project with optional versions
}
