% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_docker_compose.R
\name{generate_docker_compose}
\alias{generate_docker_compose}
\title{Generate docker-compose.yml for the project}
\usage{
generate_docker_compose(
  project_dir = getwd(),
  service_name = "app",
  image_name = "my_image"
)
}
\arguments{
\item{project_dir}{The path to the project directory (assumed to be the current working directory)}

\item{service_name}{The name of the service (default is "app")}

\item{image_name}{The name of the Docker image (default is "my_image")}
}
\description{
Generate docker-compose.yml for the project
}
