% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dockerfile.R
\name{generate_dockerfile}
\alias{generate_dockerfile}
\title{Generate Dockerfile for the project}
\usage{
generate_dockerfile(
  project_dir = getwd(),
  additional_cmds = NULL,
  use_latest_versions = FALSE
)
}
\arguments{
\item{project_dir}{The path to the project directory (assumed to be the current working directory)}

\item{additional_cmds}{Additional Dockerfile commands to be added (optional)}

\item{use_latest_versions}{Logical indicating whether to use the latest versions}
}
\description{
Generate Dockerfile for the project
}
