% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_package_versions}
\alias{get_package_versions}
\title{Get the versions of installed packages}
\usage{
get_package_versions(packages)
}
\arguments{
\item{packages}{A character vector of package names}
}
\value{
A named vector of package versions
}
\description{
Get the versions of installed packages
}
