% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_docker_environment.R
\name{setup_docker_environment}
\alias{setup_docker_environment}
\title{Setup Docker environment for the project}
\usage{
setup_docker_environment(
  project_dir = getwd(),
  use_latest_versions = FALSE,
  r_version = "latest",
  platform = "linux/amd64",
  service_name = basename(normalizePath(project_dir))
)
}
\arguments{
\item{project_dir}{The path to the project directory (assumed to be the current working directory)}

\item{use_latest_versions}{Logical indicating whether to use the latest versions}

\item{r_version}{The R version to be used in the Docker image (default is "latest")}

\item{platform}{The platform to be used (default is "linux/amd64")}

\item{service_name}{The name of the service (default is the project directory name)}
}
\description{
Setup Docker environment for the project
}
